[HUDSON-2605] Added a probe point to see why the war exploding is failing. Can you try the latest workspace again, and report the new error message?
git-svn-id: https://hudson.dev.java.net/svn/hudson/trunk/hudson/main@13278 71c3de6d-444a-0410-be80-ed276b4c234a

+++ test/src/main/java/org/jvnet/hudson/test/WarExploder.java
@@ -5,6 +5,7 @@ import hudson.FilePath;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.io.IOException;
 import java.net.URL;
 
 /**
@@ -68,6 +69,8 @@ final class WarExploder {
             System.out.println("Exploding hudson.war at "+war);
             new FilePath(explodeDir).deleteRecursive();
             new FilePath(war).unzip(new FilePath(explodeDir));
+            if(!explodeDir.exists())    // this is supposed to be impossible, but I'm investigating HUDSON-2605
+                throw new IOException("Failed to explode "+war);
             new FileOutputStream(timestamp).close();
             timestamp.setLastModified(war.lastModified());
         } else {
